feat: setup multiple patterns of test coverage
